feat: support /v1/users/{username}/following
